requirements.txt
httpx>=0.27.0
cryptography>=42.0.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.16.0
google-auth>=2.29.0
